+++ cmd/config.go
@@ -10,6 +10,7 @@ import (
 	jsoniter "github.com/json-iterator/go"
 
 	"github.com/icon-project/centralized-relay/relayer/chains/solana"
+	"github.com/icon-project/centralized-relay/relayer/chains/steller"
 	"github.com/icon-project/centralized-relay/relayer/chains/sui"
 	"github.com/icon-project/centralized-relay/relayer/chains/wasm"
 
@@ -261,6 +262,8 @@ func (iw *ProviderConfigYAMLWrapper) UnmarshalYAML(n *yaml.Node) error {
 		iw.Value = new(wasm.Config)
 	case "solana":
 		iw.Value = new(solana.Config)
+	case "stellar":
+		iw.Value = new(steller.Config)
 	case "sui":
 		iw.Value = new(sui.Config)
 	default: